로그인 과정 간소화
/login을 거치지 않고 바로 Google 로그인 페이지로 넘어가게 수정
